% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_threat.R
\name{calculate_threat}
\alias{calculate_threat}
\title{Calculating xT for passes, carries, etc}
\usage{
calculate_threat(eventData, dataType = "opta")
}
\arguments{
\item{eventData}{The dataframe that stores your data. Must contain starting x,y locations and ending x,y locations: x,y,finalX,finalY}

\item{dataType}{indicator for what type of data the eventData. Currently, options include "opta" (default) and "statsbomb"}
}
\value{
returns a dataframe object
}
\description{
Calculating xT for passes, carries, etc
}
\examples{
\dontrun{
endResult <- calculate_threat(test, dataType = "statsbomb", x_col = "location.x",
                              y_col = "location.y", xend_col = "pass.end_location.x",
                              yend_col = "pass.end_location.y")
endResult
}
}
