% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epv_guide.R
\docType{data}
\name{EPVGrid}
\alias{EPVGrid}
\title{EPV calculation data for analysis}
\format{
A \code{dataframe} with 50 columns and 31 rows spanning
a football pitch.
}
\usage{
EPVGrid
}
\description{
a simple \code{dataframe} that is used for
calculating expected possession values
}
\keyword{datasets}
