% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pass.R
\name{plot_pass}
\alias{plot_pass}
\title{Plotting passes}
\usage{
plot_pass(
  data,
  data_type = "statsbomb",
  type = "sep",
  progressive_pass = FALSE,
  cross = FALSE,
  shot = FALSE,
  switch = FALSE,
  outcome = "all",
  theme = "dark"
)
}
\arguments{
\item{data}{The data frame that stores your passing data. Opta data frame must contain atleast the following columns: \code{x}, \code{y}, \code{finalX}, \code{finalY}}

\item{data_type}{Type of data that is being put in: opta or statsbomb. Default set to "statsbomb"}

\item{type}{indicates the type of plot to pass. "sep" separates successful and unsuccessful passes. "all" plots all passes on one pitch. Default = "sep"}

\item{progressive_pass}{indicates whether to map out progressive passes}

\item{cross}{indicates whether to map out crosses}

\item{shot}{indicates whether to map out shot assists}

\item{switch}{indicates whether to map out switches of play}

\item{outcome}{indicates whether you want successful ("suc"), unsuccessful ("unsuc"), or all ("all")}

\item{theme}{indicates what theme the map must be shown in: dark (default), white, rose, almond}
}
\value{
returns a ggplot2 object
}
\description{
This function allows you to plot various types of plots
that have passes as some sort of input. Data entered must have columns for which you want to plot with.
Compatible with StatsBomb and Opta data.
}
\examples{
\dontrun{
plot  <- plot_pass(data, type = "all", progressive_pass = TRUE)
plot
}
}
