% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_passflow.R
\name{plot_passflow}
\alias{plot_passflow}
\title{Passflow plot function}
\usage{
plot_passflow(data, data_type = "statsbomb", binwidth = 0)
}
\arguments{
\item{data}{Dataframe that must house pass data only and must contain atleast the following columns: \code{x}, \code{y}, \code{finalX}, \code{finalY}}

\item{data_type}{Type of data that is being put in: opta or statsbomb. Default set to "statsbomb"}

\item{binwidth}{Details the bin size the passflow needs to bin to. The same argument name as the underlying call to \code{geom_bin2d()}. Default is 20.}
}
\value{
returns a ggplot2 object
}
\description{
This function takes in a dataframe and binsizes
to make a passflow map. Compatible, for right now, with StatsBomb data only. Returns a ggplot object
}
\examples{
\dontrun{
plot <- plot_passflow(data, binwidth = 30)
plot
}
}
