% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pizza.R
\name{plot_pizza}
\alias{plot_pizza}
\title{Function for plotting percentile/pizza plots}
\usage{
plot_pizza(
  data,
  type = "",
  template = "",
  colour_poss,
  colour_att,
  colour_def,
  player_1,
  player_2,
  colour_1,
  theme = ""
)
}
\arguments{
\item{data}{that houses data. Dataframe can contain either one player or two depending on the type of plot made.}

\item{type}{Type of plot -> single and comparison.}

\item{template}{for selecting a group of pre-selected metrics for each position by position namely:
forward, winger, midfielder, defender, goalkeeper and full back.}

\item{colour_poss}{for selecting the colour for possession group of stats. To be used only for single player plot.}

\item{colour_att}{for selecting the colour for attacking group of stats. To be used only for single player plot.}

\item{colour_def}{for selecting the colour for defense group of stats. To be used only for single player plot.}

\item{player_1}{for selecting the first player. To be used only for comparison plot.}

\item{player_2}{for selecting the second player. To be used only for comparison plot.}

\item{colour_1}{for selecting the colour of first player's bars. To be used only for comparison plot.}

\item{theme}{Theme preferences for display: dark (default), white, black.}
}
\value{
a ggplot2 object
}
\description{
This function allows for data, that has to be scraped from FBref, to be used
for plotting single and comparison percentile plots.
}
\details{
For best image quality use:
ggsave("image.png", width = 2500, height = 2800, units = "px")
}
\examples{
plot1 = plot_pizza(data = data, type = "comparison", template = "midfielder",
                             player_1 = "Nicolo Barella", player_2 = "Ilkay Gundogan", 
                             colour_1 = "lightgreen", theme = "black")

          plot2 = plot_pizza(data = data1, type = "single", template = "midfielder", 
                             colour_poss = "green", colour_att = "lightblue", 
                             colour_def = "red", theme = "dark")
}
