% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plotting scatter plots}
\usage{
plot_scatter(
  data,
  x = "",
  y = "",
  label = "",
  set_size_num = 5,
  set_size_var = "",
  set_color_num = "red",
  set_color_var = "",
  theme = "classic",
  title = "",
  title_size = 25,
  subtitle = "",
  subtitle_size = 15,
  caption = "",
  caption_size = 10
)
}
\arguments{
\item{data}{the dataframe passed in for plotting.}

\item{x}{name of column name in data to be used on x-axis}

\item{y}{name of column name in data to be used on y-axis}

\item{label}{the name of column name in data to label the scatter plot}

\item{set_size_num}{sets the size of the points set as a constant. Default size = 5.}

\item{set_size_var}{Enter name of column name in data to set size based on variable.}

\item{set_color_num}{sets the color of the points set as a constant. Can enter hexcode or a valid ggplot2 color. Default = "red"}

\item{set_color_var}{Enter name of column name in data to set color based on variable.}

\item{theme}{decide the theme of the plot between four choices: classic, minimal, grey, bw. Default = "classic"}

\item{title}{pick the title of the scatter plot}

\item{title_size}{sets the size of the title of the scatter plot. Default size = 25.}

\item{subtitle}{pick the subtitle of the scatter plot}

\item{subtitle_size}{sets the size of the subtitle of the scatter plot Default size = 15.}

\item{caption}{pick the caption of the scatter plot}

\item{caption_size}{sets the size of the caption of the scatter plot. Default size = 10.}
}
\value{
returns a ggplot2 object
}
\description{
This function allows you to plot various types of plots
that have passes as some sort of input. Compatible with any data frame of any data type. Returns a ggplot object.
}
\examples{
\dontrun{
plot <- plot_scatter(data, x = "player", y = "age", label = "team")
plot
}
}
