% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shot.R
\name{plot_shot}
\alias{plot_shot}
\title{Function for plotting shots}
\usage{
plot_shot(
  data,
  type = "",
  bins = 30,
  highlight_goals = "",
  average_location = "",
  theme = ""
)
}
\arguments{
\item{data}{Dataframe that houses shot data. Dataframe must contain atleast the following columns: X,Y,xG,result,name}

\item{type}{Type of showcasing the shotmap: hexbin, density, point (default)}

\item{bins}{Bin size for creating bins. Use this when using hexbin shotmap. Default = 30.}

\item{highlight_goals}{to choose to display only the goals in a different color.}

\item{average_location}{for removing lines denoting average location of shots if need be.}

\item{theme}{Theme preferences for display: dark (default), white, rose, almond}
}
\value{
a ggplot2 object
}
\description{
This function allows for data, that has to be scraped from Understat, to be used
for plotting shots.
}
\examples{
\dontrun{
plot <- plot_shot(data, type = "hexbin", bins = 20, average_location = TRUE, highlight_goals = FALSE)
}
}
