% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sonar.R
\name{plot_sonar}
\alias{plot_sonar}
\title{Function for plotting pass sonars}
\usage{
plot_sonar(data, data_type = "statsbomb", title = "")
}
\arguments{
\item{data}{Dataframe that houses pass data. Dataframe must contain atleast the following columns: \code{x}, \code{y}, \code{finalX}, \code{finalY}}

\item{data_type}{Type of data that is being put in: opta or statsbomb. Default set to "statsbomb"}

\item{title}{Title of the passing sonar plot}
}
\value{
a ggplot2 object
}
\description{
This function allows for data, that can be from Opta or Statsbomb, to be used
for plotting pass sonars.
}
\examples{
\dontrun{
plot <- plot_sonar(data, data_type = "statsbomb")
plot
}
}
