% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeline.r
\name{plot_timeline}
\alias{plot_timeline}
\title{Create timeline charts (xG) using event data.}
\usage{
plot_timeline(
  data,
  match_year,
  team_home,
  team_away,
  color_home,
  color_away,
  theme = ""
)
}
\arguments{
\item{data}{is for the dataset used.}

\item{match_year}{the year the match was played.}

\item{team_home}{is for the home team according to data.}

\item{team_away}{is for the away team according to data.}

\item{color_home}{is for the color of the line for the home team.}

\item{color_away}{is for the color of the line for the away team.}

\item{theme}{to select the colors. Choose from 4 themes -> dark, almond, rose, white.}
}
\value{
a ggplot2 object
}
\description{
This function allows to make match timelines using data collected event-by-event.
}
\examples{
\dontrun{
plot <- plot_timeline(data = data, match_year = 2021, 
                      team_home = "Manchester United", team_away = "Liverpool",
                      color_home = "#e31a1c", color_away = "#980043", theme = "dark")
plot
}
}
