% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_event_data_guide.R
\docType{data}
\name{SampleEventData}
\alias{SampleEventData}
\title{Sample event data frame}
\format{
A \code{dataframe} with the columns \code{x}, \code{y}, \code{finalX}, \code{finalY}, \code{minute},
\code{playerId}, \code{teamId}, \code{outcome} and \code{type}.
}
\usage{
SampleEventData
}
\description{
Event \code{dataframe} to play with and understand usage of functions
}
\keyword{datasets}
