% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sbevent_data_guide.R
\docType{data}
\name{SampleSBData}
\alias{SampleSBData}
\title{Sample StatsBomb event data frame}
\format{
A \code{dataframe} with the columns \code{x}, \code{y}, \code{finalX}, \code{finalY}, and the rest of StatsBomb columns
}
\usage{
SampleSBData
}
\description{
StatsBomb event \code{dataframe} to play with and understand usage of functions.
The following code was used with the use of {StatsBombR}:

\code{ Comp <- FreeCompetitions() \%>\% }
\code{ filter(competition_id == 11 & season_name == "2014/2015") }
\code{ Matches <- FreeMatches(Comp) }
\code{ StatsBombData <- StatsBombFreeEvents(MatchesDF = Matches, Parallel = TRUE) }
\code{ plotting_data  <- allclean(StatsBombData) }
\code{ plotting_data <- plotting_data \%>\% }
\code{ rename("x" = "location.x", }
\code{ "y" = "location.y", }
\code{ "finalX" = "pass.end_location.x", }
\code{ "finalY" = "pass.end_location.y") }
}
\keyword{datasets}
