% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shift_column}
\alias{shift_column}
\title{shift_column}
\usage{
shift_column(
  data,
  columns,
  new_names = sprintf("\%s.Shifted", columns),
  len = 1L,
  up = TRUE
)
}
\arguments{
\item{data}{data}

\item{columns}{columns to shift over}

\item{new_names}{New name of shifted columns, Default: sprintf("\%s.Shifted", columns)}

\item{len}{length of rows to shift, Default: 1}

\item{up}{shift rows up (TRUE) or down (FALSE), Default: TRUE}
}
\value{
data.frame with shifted columns
}
\description{
Port of \code{shift.column()} from Jared Lander's {useful} package
with a few minor changes.
}
\keyword{internal}
