% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{viridis_d_pal}
\alias{viridis_d_pal}
\title{viridis_d_pal}
\format{
An object of class \code{character} of length 256.
}
\usage{
viridis_d_pal
}
\description{
Viridis palette hex codes. Used in \code{plot_passflow()} function.
Created by Stefan van der Walt & Nathaniel Smith for matplotlib.
}
\keyword{internal}
