% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threat_guide.R
\docType{data}
\name{xTGrid}
\alias{xTGrid}
\title{xT calculation data for analysis}
\format{
A \code{dataframe} with 12 columns and 8 rows spanning
a football pitch.
}
\usage{
xTGrid
}
\description{
a simple \code{dataframe} that is used for
calculating expected threat values
}
\keyword{datasets}
