% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{zissou_pal}
\alias{zissou_pal}
\title{title zissou_pal}
\format{
An object of class \code{character} of length 10.
}
\usage{
zissou_pal
}
\description{
'Zissou' palette hex codes. Useed in \code{plot_sonar()} function.
Created by Karthik Ram in the {wesanderson} R package.
}
\keyword{internal}
