% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_threat.R
\name{calculate_threat}
\alias{calculate_threat}
\title{Calculating xT for passes, carries, etc}
\usage{
calculate_threat(data, type = "opta")
}
\arguments{
\item{data}{The dataframe that stores your data. Must contain starting x,y locations and ending x,y locations: \code{x}, \code{y}, \code{finalX}, \code{finalY}}

\item{type}{indicator for what type of data the data. Currently, options include "opta" (default) and "statsbomb"}
}
\value{
returns a dataframe object
}
\description{
Calculating xT for passes, carries, etc
}
\examples{
\dontrun{
endResult <- calculate_threat(test, type = "statsbomb")
endResult
}
}
