% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_convexhull.R
\name{plot_convexhull}
\alias{plot_convexhull}
\title{Function for plotting convex hulls}
\usage{
plot_convexhull(
  data,
  data_type = "statsbomb",
  color = "#E74C3C",
  title = "",
  theme = "dark"
)
}
\arguments{
\item{data}{Data frame that houses pass data. Opta dataframe must contain atleast the following columns: \code{x}, \code{y}, \code{finalX}, \code{finalY}, \code{playerId}}

\item{data_type}{Type of data that is being put in: opta or statsbomb. Default set to "statsbomb"}

\item{color}{The color of the outline of the convex hull}

\item{title}{Title of the plot}

\item{theme}{Indicates what theme the map must be shown in: dark (default), white, rose, almond}
}
\value{
a ggplot2 object
}
\description{
This function allows for data, that can be from Opta or StatsBomb, to be used
for plotting convex hulls on top of an outline of a football pitch.
}
\examples{
\dontrun{
plot <- plot_convexhull(data = data, data_type = "opta", color = "blue", title = "Team 1")
plot
}
}
