% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plotting heatmap}
\usage{
plot_heatmap(eventData, type = "", theme = "")
}
\arguments{
\item{eventData}{The dataframe that stores your data. Must contain starting x,y locations}

\item{type}{indicates the type of heatmap to plot. "hex" indicates hex bins, "density" indicates density (default), and
"bin" indicates bin heatmap pass}

\item{theme}{indicates what theme the map must be shown in: dark (default), white, rose, almond}
}
\value{
returns a ggplot2 object
}
\description{
This function allows you to plot various types of heatmaps of starting x and y coordinates:
hex bin heatmap,
density heatmap, and
bin heatmap
}
\examples{
\dontrun{
plot <- plot_heatmap(touchData, type="hex")
plot
}
}
