% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_passnet.R
\name{plot_passnet}
\alias{plot_passnet}
\title{Function for plotting pass networks}
\usage{
plot_passnet(
  data,
  data_type = "statsbomb",
  team_name,
  color_scale = "#E74C3C",
  subtitle_plot = "",
  theme = "dark"
)
}
\arguments{
\item{data}{Data frame that houses pass data}

\item{data_type}{Type of data that is being put in: opta or statsbomb. Default set to "statsbomb"}

\item{team_name}{The name of the team of which you want a pass network}

\item{color_scale}{Color of higher end of xT color scale. Default set to "#E74C3C"}

\item{subtitle_plot}{Subtitle of the pass network plot}

\item{theme}{The background theme -> "light" or "dark"}
}
\value{
a ggplot2 object
}
\description{
This function allows for data, that can be from Opta or Statsbomb, to be used
for plotting pass networks.
}
\examples{
\dontrun{
plot <- plot_passnet(data, data_type = "opta", team_name = "Arsenal")
plot
}
}
