% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pizza.R
\name{plot_pizza}
\alias{plot_pizza}
\title{Function for plotting percentile/pizza plots}
\usage{
plot_pizza(
  data,
  type = "",
  template = "",
  colour_poss,
  colour_att,
  colour_def,
  player_1,
  player_2,
  colour_compare,
  season,
  season_player_1,
  season_player_2,
  theme = ""
)
}
\arguments{
\item{data}{that houses data. Dataframe can contain either one player or two depending on the type of plot made.}

\item{type}{Type of plot -> single and comparison.}

\item{template}{for selecting a group of pre-selected metrics for each position by position namely:
forward, winger, midfielder, defender, goalkeeper and full back.}

\item{colour_poss}{for selecting the colour for possession group of stats. To be used only for single player plot.}

\item{colour_att}{for selecting the colour for attacking group of stats. To be used only for single player plot.}

\item{colour_def}{for selecting the colour for defense group of stats. To be used only for single player plot.}

\item{player_1}{for selecting the first player. To be used only for comparison plot.}

\item{player_2}{for selecting the second player. To be used only for comparison plot.}

\item{colour_compare}{for selecting the colour of comparison to be used only for comparison plot.}

\item{season}{Specify what season to pick for a single player pizza chart. Pick the scouting period from the scouting period column in the data.}

\item{season_player_1}{Specify what season to pick for the first player in a pizza chart}

\item{season_player_2}{Specify what season to pick for the second player in a pizza chart}

\item{theme}{Specify the theme of the pizza chart -> dark, black, and white. Default set to black}
}
\value{
a ggplot2 object
}
\description{
This function allows for data, that has to be scraped from FBref, to be used
for plotting single and comparison percentile plots.
}
\details{
For best image quality use:
ggsave("image.png", width = 2900, height = 2800, units = "px")
}
\examples{
\dontrun{
plot1 <- plot_pizza(data = data, type = "comparison", template = "midfielder",
                    player_1 = "Nicolo Barella", player_2 = "Ilkay Gundogan",
                    season_player_1 = "Last 365 Days", season_player_2 = "Last 365 Days",
                    colour_compare = "lightgreen", theme = "black")
plot1

plot2 <- plot_pizza(data = data1, type = "single", template = "midfielder",
                    colour_poss = "green", colour_att = "lightblue", season = "Last 365 Days",
                    colour_def = "red", theme = "dark")
plot2
}
}
