% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plotting scatter plots}
\usage{
plot_scatter(
  data,
  scatter_x = "",
  scatter_y = "",
  sc_label = "",
  set_size_num = 5,
  set_size_var = "",
  set_color_num = "red",
  set_color_var = "",
  theme = "classic",
  scatter_title = "",
  title_size = 25,
  scatter_subtitle = "",
  subt_size = 15,
  scatter_cap = "",
  cap_size = 10
)
}
\arguments{
\item{data}{the dataframe passed in for plotting.}

\item{scatter_x}{name of column name in data to be used on x-axis}

\item{scatter_y}{name of column name in data to be used on y-axis}

\item{sc_label}{the name of column name in data to be label the scatter plot}

\item{set_size_num}{sets the size of the points set as a constant. Default size = 5.}

\item{set_size_var}{Enter name of column name in data to set size based on variable.}

\item{set_color_num}{sets the color of the points set as a constant. Can enter hexcode or a valid ggplot2 color. Default = "red"}

\item{set_color_var}{Enter name of column name in data to set color based on variable.}

\item{theme}{decide the theme of the plot between four choices: classic, minimal, grey, bw. Default = "classic"}

\item{scatter_title}{pick the title of the scatter plot}

\item{title_size}{sets the size of the title of the scatter plot. Default size = 25.}

\item{scatter_subtitle}{pick the subtitle of the scatter plot}

\item{subt_size}{sets the size of the subtitle of the scatter plot Default size =15.}

\item{scatter_cap}{pick the caption of the scatter plot}

\item{cap_size}{sets the size of the caption of the scatter plot. Default size = 10.}
}
\value{
returns a ggplot2 object
}
\description{
This function allows you to plot various types of plots
that have passes as some sort of input. Compatible with any data frame of any data type. Returns a ggplot object.
}
\examples{
\dontrun{
plot <- plot_scatter(dtaa, scatter_x = "player", scatter_y = "age", sc_label = "team")
plot
}
}
