% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sonar.R
\name{plot_sonar}
\alias{plot_sonar}
\title{Function for plotting pass sonars}
\usage{
plot_sonar(sonarData, dataType = "statsbomb", titlePlot = "")
}
\arguments{
\item{sonarData}{Dataframe that houses pass data. Dataframe must contain atleast the following columns: x,y,finalX,finalY}

\item{dataType}{Type of data that is being put in: opta or statsbomb. Default set to "statsbomb"}

\item{titlePlot}{Title of the passing sonar plot}
}
\value{
a ggplot2 object
}
\description{
This function allows for data, that can be from Opta or Statsbomb, to be used
for plotting pass sonars.
}
\examples{
plot = plot_sonar(sonarData, dataType="statsbomb")
}
