% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeline.R
\name{plot_timeline}
\alias{plot_timeline}
\title{Create timeline charts (xG) using 'understat' event data.}
\usage{
plot_timeline(
  data,
  match_year,
  team_home,
  team_away,
  home_color,
  away_color,
  theme = ""
)
}
\arguments{
\item{data}{Data, for now only compatible with understat data.}

\item{match_year}{the year the match was played.}

\item{team_home}{home team according to data.}

\item{team_away}{away team according to data.}

\item{home_color}{line colour for the home team.}

\item{away_color}{line colour for the away team.}

\item{theme}{Choose from 4 ggplot2 themes -> dark, almond, rose, white.}
}
\value{
ggplot2 object of a xG timeline plot
}
\description{
This function allows to make match timelines using data collected event-by-event.
}
\examples{
\dontrun{
plot <- plot_timeline(data = data, match_year = 2021, 
                      team_home = "Manchester United", team_away = "Liverpool",
                      home_color = "#e31a1c", away_color = "#980043", theme = "dark")
plot
}
}
