% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeline.r
\name{plot_timeline}
\alias{plot_timeline}
\title{Create timeline charts (xG) using event data.}
\usage{
plot_timeline(
  data,
  match_year,
  team_home,
  team_away,
  home_color,
  away_color,
  theme = ""
)
}
\arguments{
\item{data}{is for the dataset used.}

\item{team_home}{is for the home team according to data.}

\item{team_away}{is for the away team according to data.}

\item{home_color}{is for the colour of the line for the home team.}

\item{away_color}{is for the colour of the line for the away team.}

\item{theme}{to select the colours. Choose from 4 themes -> dark, almond, rose, white.}
}
\value{
a ggplot2 object
}
\description{
This function allows to make match timelines using data collected event-by-event.
}
\examples{
\dontrun{
plot <- plot_timeline(data = data, match_year = 2021, team_home = "Manchester United", team_away = "Liverpool",
                      home_color = "#e31a1c", away_color = "#980043", theme = "dark")
plot
}
}
