% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeline.r
\name{plot_timeline}
\alias{plot_timeline}
\title{Create timeline charts (xG) using event data.}
\usage{
plot_timeline(
  data,
  match_year,
  team_home,
  team_away,
  home_color,
  away_color,
  theme = ""
)
}
\arguments{
\item{data}{Data frame used in the function}

\item{match_year}{Year the match was played}

\item{team_home}{Home team according to data}

\item{team_away}{Away team according to data}

\item{home_color}{Color of the line for the home team}

\item{away_color}{Color of the line for the away team}

\item{theme}{to select the colors. Choose from 4 themes -> dark, almond, rose, white}
}
\value{
a ggplot2 object
}
\description{
This function allows to make match timelines using data collected event-by-event.
}
\examples{
\dontrun{
plot <- plot_timeline(data = data, match_year = 2021, 
                      team_home = "Manchester United", team_away = "Liverpool",
                      home_color = "#e31a1c", away_color = "#980043", theme = "dark")
plot
}
}
