% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_voronoi.R
\name{plot_voronoi}
\alias{plot_voronoi}
\title{Function for plotting voronoi diagrams on the pitch}
\usage{
plot_voronoi(
  data,
  data_type = "statsbomb",
  colour = "#E74C3C",
  fill = "",
  alpha = 0.4,
  title = "",
  theme = "dark"
)
}
\arguments{
\item{data}{Dataframe that houses pass data. Opta/Statsbomb dataframe must contain atleast the following columns: \code{x}, \code{y}}

\item{data_type}{Type of data that is being put in: opta or statsbomb. Default set to "statsbomb".}

\item{colour}{The colour of the points in the voronoi plot.}

\item{fill}{Name of column to add a fill component to the plot}

\item{alpha}{Alpha value for opacity of fill. Default set to 0.4}

\item{title}{Title of the plot.}

\item{theme}{Indicates what theme the map must be shown in: dark (default), white, rose, almond.}
}
\value{
a ggplot2 object
}
\description{
The function allows for creation of voronoi diagrams on a football pitch
with either Opta or Statsbomb data.
}
\examples{
\dontrun{
plot <- plot_voronoi(data = data, colour = "blue", title = "Team 1")
plot
}

}
