% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sbevent_data_guide.R
\docType{data}
\name{SampleSBData}
\alias{SampleSBData}
\title{Sample StatsBomb event data frame}
\format{
A \code{dataframe} with the columns \code{x}, \code{y}, \code{finalX}, \code{finalY}, and the rest of StatsBomb columns
}
\usage{
SampleSBData
}
\description{
StatsBombe event \code{dataframe} to play with and understand usage of functions.
The following code was used with the use of {StatsBombR}:
Comp <- FreeCompetitions() \%>\%
filter(competition_id == 11 & season_name == "2014/2015")
Matches <- FreeMatches(Comp)
StatsBombData <- StatsBombFreeEvents(MatchesDF = Matches, Parallel = TRUE)
plotting_data  <- allclean(StatsBombData)
plotting_data <- plotting_data \%>\%
rename("x" = "location.x",
"y" = "location.y",
"finalX" = "pass.end_location.x",
"finalY" = "pass.end_location.y")
}
\keyword{datasets}
